+++ CHANGELOG.md
@@ -5,12 +5,18 @@
 - `--editor` flag for `doing last` to edit last entry
 - `--tag=` flag to filter `doing last` by tag
 - `--search=` to filter `doing last` by text/regex search
+- `--search=` for `doing finish`, finish last X entries matching search
 - Add `tags_color` as a primary config key to highlight @tags in displayed entries
 
 #### IMPROVED
 
+- Code cleanup
 - Clean up command line help
 - --editor improvements for all commands that use it
+
+#### FIXED
+
+- `doing tag --remove` was adding tags if they didn't exist
 
 ### 1.0.54